even more trim
